Fixed issue #217
Bug: integral division result cast to double or float
Pattern id: ICAST_IDIV_CAST_TO_DOUBLE, type: ICAST, category: STYLE


This code casts the result of an integral division (e.g., int or long division) operation to double or float. Doing division on integers truncates the result to the integer value closest to zero. The fact that the result was cast to double suggests that this precision should have been retained. What was probably meant was to cast one or both of the operands to double before performing the division. Here is an example: 

int x = 2;
int y = 5;
// Wrong: yields result 0.0
double value1 =  x / y;

// Right: yields result 0.4
double value2 =  x / (double) y;

+++ trunk/EvolutionChamber/src/java/com/fray/evo/EvolutionChamber.java
@@ -341,7 +341,7 @@ public final class EvolutionChamber
 			{
 				Collections.shuffle(conf.getGeneticOperators());
 				BASE_MUTATION_RATE += .001;
-				if (BASE_MUTATION_RATE >= CHROMOSOME_LENGTH / 2)
+				if (BASE_MUTATION_RATE >= (double) CHROMOSOME_LENGTH / 2.0 )
 					BASE_MUTATION_RATE = 1;
 				IChromosome fittestChromosome = population.getFittestChromosome();
 				if (killThreads)
